added gauss label and added 2c interpretation to temp output

+++ main/drivers/include/mag.h
@@ -15,7 +15,7 @@ uint16_t get_y_output_raw(i2c_inst_t *i2c);
 int16_t get_z_output(i2c_inst_t *i2c);
 uint16_t get_z_output_raw(i2c_inst_t *i2c);
 
-int get_temp_output(i2c_inst_t *i2c);
+int16_t get_temp_output(i2c_inst_t *i2c);
 uint16_t get_temp_output_raw(i2c_inst_t *i2c);
 
 uint8_t get_mag_status(i2c_inst_t *i2c);
